+++ internal/service/loadbalancer/backend.go
@@ -3,6 +3,7 @@ package loadbalancer
 import (
 	"context"
 
+	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/request"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/service"
@@ -101,7 +102,7 @@ func resourceBackendRead(ctx context.Context, d *schema.ResourceData, meta inter
 		Name:        name,
 	})
 	if err != nil {
-		return handleResourceError(d.Get("name").(string), d, err)
+		return utils.HandleResourceError(d.Get("name").(string), d, err)
 	}
 
 	d.SetId(marshalID(serviceID, be.Name))