+++ internal/service/loadbalancer/dynamic_certificate_bundle.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"time"
 
+	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/request"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/service"
@@ -99,7 +100,7 @@ func resourceDynamicCertificateBundleRead(ctx context.Context, d *schema.Resourc
 		UUID: d.Id(),
 	})
 	if err != nil {
-		return handleResourceError(d.Get("name").(string), d, err)
+		return utils.HandleResourceError(d.Get("name").(string), d, err)
 	}
 
 	if diags = setDynamicCertificateBundleResourceData(d, b); len(diags) > 0 {