+++ internal/service/loadbalancer/frontend_rule.go
@@ -3,6 +3,7 @@ package loadbalancer
 import (
 	"context"
 
+	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/request"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/service"
@@ -124,7 +125,7 @@ func resourceFrontendRuleRead(ctx context.Context, d *schema.ResourceData, meta 
 		Name:         name,
 	})
 	if err != nil {
-		return handleResourceError(d.Get("name").(string), d, err)
+		return utils.HandleResourceError(d.Get("name").(string), d, err)
 	}
 
 	d.SetId(marshalID(serviceID, feName, rule.Name))