+++ internal/service/loadbalancer/frontend_tls_config.go
@@ -3,6 +3,7 @@ package loadbalancer
 import (
 	"context"
 
+	"github.com/UpCloudLtd/terraform-provider-upcloud/internal/utils"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/request"
 	"github.com/UpCloudLtd/upcloud-go-api/v5/upcloud/service"
@@ -83,7 +84,7 @@ func resourceFrontendTLSConfigRead(ctx context.Context, d *schema.ResourceData, 
 		Name:         name,
 	})
 	if err != nil {
-		return handleResourceError(d.Get("name").(string), d, err)
+		return utils.HandleResourceError(d.Get("name").(string), d, err)
 	}
 
 	if err = d.Set("frontend", marshalID(serviceID, feName)); err != nil {